+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Changelog
 
 - dropped support for Python 3.7, added support for Python 3.11
 
+### Added
+
+- added configurable socket timeout for requests to the Apify API
+
 ### Internal changes
 
 - simplified retrying with exponential backoff